py-env/pipeline/requirements.txt
requests
kafka-python
hdfs